cursus t.e.m. p167

+++ minicms-pro/admin/index.php
@@ -2,11 +2,20 @@
 declare(strict_types=1);
 require __DIR__ . '/autoload.php';
 
+/**
+ * Start de PHP session.
+ * Zonder dit werkt $_SESSION niet.
+ */
+session_start();
+
 use Admin\Controllers\DashboardController;
 use Admin\Controllers\PostsController;
 use Admin\Controllers\ErrorController;
+use Admin\Controllers\AuthController;
 use Admin\Core\Router;
+use Admin\Core\Auth;
 use Admin\Repositories\PostsRepository;
+use Admin\Repositories\UsersRepository;
 use Admin\Models\StatsModel;
 
 $uri = parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH);
@@ -19,6 +28,12 @@ if (str_starts_with($uri, $basePath)) {
 $uri = rtrim($uri, '/');
 $uri = $uri === '' ? '/' : $uri;
 
+$publicRoutes = ['/login'];
+if (!Auth::check() && !in_array($uri, $publicRoutes, true)) {
+    header('Location: /php/minicms-pro/admin/login');
+    exit;
+}
+
 $method = $_SERVER['REQUEST_METHOD'];
 
 $router = new Router();
@@ -27,8 +42,7 @@ $router = new Router();
  * setNotFoundHandler()
  *
  * Doel:
- * - Zorgt dat elke onbekende URL een nette 404 pagina krijgt via
-ErrorController.
+ * - Zorgt dat elke onbekende URL een nette 404 pagina krijgt via ErrorController.
  */
 $errorController = new ErrorController();
 $router->setNotFoundHandler(function (string $requestedUri) use
@@ -37,14 +51,36 @@ $router->setNotFoundHandler(function (string $requestedUri) use
 });
 
 /**
+ * Auth
+ */
+$router->get('/login', function (): void {
+    (new AuthController(UsersRepository::make()))->showLogin();
+});
+$router->post('/login', function (): void {
+    (new AuthController(UsersRepository::make()))->login();
+});
+$router->post('/logout', function (): void {
+    (new AuthController(UsersRepository::make()))->logout();
+});
+
+/**
  * Delete (confirm + action)
  */
 $router->get('/posts/{id}/delete', function (int $id): void {
+    if (!Auth::isAdmin()) {
+        header('Location: /php/minicms-pro/admin/posts');
+        exit;
+    }
     (new PostsController(PostsRepository::make()))->deleteConfirm($id);
 });
 $router->post('/posts/{id}/delete', function (int $id): void {
+    if (!Auth::isAdmin()) {
+        header('Location: /php/minicms-pro/admin/posts');
+        exit;
+    }
     (new PostsController(PostsRepository::make()))->delete($id);
 });
+
 
 /**
  * Dashboard
